added router to app.js and deleted router file
